% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mo.R
\name{moDetail_query}
\alias{moDetail_query}
\title{查询生产订单明细信息}
\usage{
moDetail_query(
  token = "D62212AC-5651-4387-B3AC-42D3E38D6AFD",
  FStartDate = "2023-03-01",
  FEndDate = "2023-04-30",
  Fcategory = NULL
)
}
\arguments{
\item{token}{口令}

\item{FStartDate}{开始日期}

\item{FEndDate}{结束日期}
}
\value{
返回值
}
\description{
查询生产订单明细信息
}
\examples{
moDetail_query()
}
