% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mo.R
\name{moStat_query}
\alias{moStat_query}
\title{查询生产订单汇总信息}
\usage{
moStat_query(
  token = "D62212AC-5651-4387-B3AC-42D3E38D6AFD",
  FStartDate = "2023-03-01",
  FEndDate = "2023-04-30"
)
}
\arguments{
\item{token}{口令}

\item{FStartDate}{开始日期}

\item{FEndDate}{结束日期}
}
\value{
返回值
}
\description{
查询生产订单汇总信息
}
\examples{
moStat_query()
}
