% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overDueList.R
\name{overDueListPrd_query}
\alias{overDueListPrd_query}
\title{产成品逾期查询}
\usage{
overDueListPrd_query(token = "D62212AC-5651-4387-B3AC-42D3E38D6AFD")
}
\arguments{
\item{token}{口令}
}
\value{
返回值
}
\description{
产成品逾期查询
}
\examples{
overDueListPrd_query()
}
