% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overDueList.R
\name{overDuePrd_delete}
\alias{overDuePrd_delete}
\title{产品逾期表删除所有数据}
\usage{
overDuePrd_delete(token = "D62212AC-5651-4387-B3AC-42D3E38D6AFD")
}
\arguments{
\item{token}{口令}
}
\value{
返回值
}
\description{
产品逾期表删除所有数据
}
\examples{
overDuePrd_delete()
}
