% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overDueList.R
\name{overDuePrd_query}
\alias{overDuePrd_query}
\title{产成品逾期查询}
\usage{
overDuePrd_query(
  token = "D62212AC-5651-4387-B3AC-42D3E38D6AFD",
  FCategory = NULL
)
}
\arguments{
\item{token}{口令}

\item{FCategory}{产品分组}
}
\value{
返回值
}
\description{
产成品逾期查询
}
\examples{
overDueListPrd_query()
}
