% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overDueList.R
\name{overDueSParts_insert}
\alias{overDueSParts_insert}
\title{备件逾期表数据将视图数据插入对应表}
\usage{
overDueSParts_insert(token = "D62212AC-5651-4387-B3AC-42D3E38D6AFD")
}
\arguments{
\item{token}{口令}
}
\value{
返回值
}
\description{
备件逾期表数据将视图数据插入对应表
}
\examples{
overDueSParts_insert()
}
