% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overDueList.R
\name{overDueSParts_query_cn}
\alias{overDueSParts_query_cn}
\title{备件逾期表查询数据}
\usage{
overDueSParts_query_cn(token = "D62212AC-5651-4387-B3AC-42D3E38D6AFD")
}
\arguments{
\item{token}{口令}
}
\value{
返回值
}
\description{
备件逾期表查询数据
}
\examples{
overDueSParts_query_cn()
}
